backend/requirements.txt
# backend/requirements.txt - UPDATED VERSION
fastapi==0.109.2
uvicorn[standard]==0.27.0
python-dotenv==1.0.1
openai==1.12.0
pydantic==2.5.3
python-multipart==0.0.9
httpx==0.26.0
aiofiles==23.2.1
python-json-logger==2.0.7